fix: Handle floating point frequency values.
At times the cable modem returns frequency information as floating point
values instead of integers, and this causes the parsing logic to return
an error. This happens generally when the cable modem is just starting
up and hasn't yet established a connection.

The v0.2.0 version of the cablemodemutil fixes this issue, however
changes the type of all frequency related values to float32 instead of
uint32. This change handles the type change.

In addition, v0.2.0 of the cablemodemutil library contains the
downstream and upstream related information in separate sub-fields. This
change also handles the updated field structure.